testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/history_back.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 <body>
	<p>
		<h3>history_back</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		function onPopstateEvent(e) {
			test(function() {
				assert_equals(e.state, 1, "Current page state is not 1.");
			});
		}

		if (window.addEventListener) {
            window.addEventListener('popstate', onPopstateEvent, false);
        }
        else if (window.attachEvent) {
            window.attachEvent("onpopstate", onPopstateEvent);
        }
        else {
            test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			});
        }

		try {
			window.history.pushState(1, document.title, '?x=1');
			window.history.pushState(2, document.title, '?x=2');
		} catch (e) {
			test(function() {
				assert_unreached("call of pushState method is failed.");
			});
		}

		window.history.back();

	</script>
 </body>
</html>
